src/test/ClienteServiceTest.java
package br.com.betereli;

import br.com.betereli.Service.ClienteService;
import br.com.betereli.dao.ClienteDao;
import br.com.betereli.dao.ClienteDaoMock;
import br.com.betereli.dao.IClienteDao;
import org.junit.Assert;
import org.junit.Test;

public class ClienteServiceTest {

    @Test
    public void salvarTeste(){
        ClienteDaoMock mockDao = new ClienteDaoMock();
        ClienteService service = new ClienteService(mockDao);
        String retorno = service.salvar();
        Assert.assertEquals("Sucesso", retorno);
    }

    @Test(expected = UnsupportedOperationException.class)
    public void esperadoErroNosalvarTeste(){
        IClienteDao mockDao = new ClienteDao();
        ClienteService service = new ClienteService(mockDao);
        String retorno = service.salvar();
        Assert.assertEquals("Sucesso", retorno);
    }
}
